+++ utilities/external-interface/external_interface.cpp
@@ -1,8 +1,6 @@
-#include "execution_context.hpp"
+#include "executor.hpp"
 #include "format_support.hpp"
 #include "processing.hpp"
-#include "processing_step.hpp"
-#include "processing_step_factory.hpp"
 #include "utils.hpp"
 
 #include <SeQuant/core/context.hpp>
@@ -653,74 +651,6 @@ void registerIndexSpaces(const json &spaces, IndexSpaceMeta &meta) {
   }
 }
 
-void process_steps(const json &json_steps, const IndexSpaceMeta &meta) {
-  (void)meta;
-  if (!json_steps.is_array()) {
-    throw Exception("Steps object must be an array");
-  }
-
-  ExecutionContext ctx;
-
-  std::size_t step_id_counter = 0;
-
-  for (const json &step : json_steps) {
-    const std::string_view kind = step.at("kind").get<std::string_view>();
-
-    std::vector<std::string_view> inputs;
-
-    if (step.contains("inputs")) {
-      const nlohmann::json &inps = step.at("inputs");
-      if (inps.is_string()) {
-        inputs.emplace_back(inps.get<std::string_view>());
-      } else if (inps.is_array()) {
-        for (const auto &current : inps) {
-          if (!current.is_string()) {
-            throw Exception("Entries in inputs array must be strings");
-          }
-
-          inputs.emplace_back(current.get<std::string_view>());
-        }
-      } else {
-        throw Exception("inputs field must be either a string or an array");
-      }
-    }
-
-    std::unique_ptr<ProcessingStep> proc_step =
-        ProcessingStepFactory::instance().instantiate(kind);
-
-    if (step.contains("options")) {
-      if (!proc_step->accepts_options()) {
-        throw Exception("Processing step '" + std::string(kind) +
-                        "' does not take options but some where given");
-      }
-
-      proc_step->set_options(step.at("options"));
-    } else if (proc_step->requires_options()) {
-      throw Exception("Processing step '" + std::string(kind) +
-                      "' requires options but none where given");
-    }
-
-    std::string step_id = step.contains("id") ? step.at("id").get<std::string>()
-                                              : std::to_string(step_id_counter);
-
-    proc_step->run(step_id, ctx, inputs);
-
-    if (step.contains("outputs")) {
-      const nlohmann::json &outputs = step.at("outputs");
-      if (!outputs.is_object()) {
-        throw Exception("outputs field must be an object");
-      }
-
-      for (const auto &[key, val] : outputs.items()) {
-        ctx.add_data_alias(step_id + "[" + val.get<std::string>() + "]",
-                           step_id + "." + key);
-      }
-    }
-
-    ++step_id_counter;
-  }
-}
-
 void process(const json &driver, IndexSpaceMeta &spaceMeta) {
   if (!driver.contains("index_spaces")) {
     throw Exception("Missing index_spaces definition");
@@ -732,7 +662,8 @@ void process(const json &driver, IndexSpaceMeta &spaceMeta) {
     throw Exception("Missing steps specification");
   }
 
-  process_steps(driver.at("steps"), spaceMeta);
+  Executor executor;
+  executor.execute(driver.at("steps"));
 }
 
 void generalSetup() {
